Bug fix for DerivedData/DG for face data-structures in parallel: For some reason an if-test was added around a year ago which bypassed genNtfac(), genInpofaTet() and genEsuf() when a chunk/partition did not have a physical boundary on them. This gave wrong results when running in parallel with many partitions, and some of these partitions ended up without a physical boundary; which is a totally normal scenario. Thanks @jbakosi for helping find this horrible bug!

+++ src/Inciter/DG.C
@@ -649,7 +649,7 @@ DG::nodetripletMatch( const std::array< std::size_t, 2 >& id,
   for (std::size_t k=0; k<4; ++k)
   {
     auto el = esuf[ 2*id[0] ];
-    auto ip = inpoel[ 4*el+k ];
+    auto ip = inpoel[ 4*static_cast< std::size_t >( el )+k ];
     Assert( el == static_cast< int >( id[1] ), "Mismatch in id and esuf" );
     for (std::size_t j=0; j<3; ++j)
     {
